Use fastify, a web framework.
See https://www.fastify.io/.

Compared to express, fastify is faster, and it seems like it is
better-regarded. It does import a few less packages too, although it is
still pretty big.

In the long term we should certainly be exporting a specific directory
for static files instead, but I didn't want to put that in this PR.